fix: warnings; Update README
