+++ README.md
@@ -1,12 +1,18 @@
 # Summa Aggregation
 
-Summa Aggregation is focused on optimizing the generation of Merkle sum tree, a task identified as the primary time-consuming process in Summa benchmarks. Our goal is to significantly reduce the time required to generate these tree by leveraging parallelization across multiple machines.
+Summa Aggregation is a scalable solution specifically designed to accelerate the process of building Merkle sum tree.
 
-The system features `AggregationMerkleSumTree`, a specialized component designed for the efficient construction of complete tree from smaller, aggregated structures known as `mini-tree`. This approach not only speeds up the tree-building process but also enhances scalability and performance in large-scale data environments.
+Our benchmarks in Summa Solvency revealed that constructing a merkle sum tree is a major time-consuming aspect of proof generation.
+
+The primary objective of Summa Aggregation is to enable efficient scaling in the construction of the tree by parallelization and distributed computation in multiple machines.
+
+For further optimization in Summa Aggregation, we introduced the `AggregationMerkleSumTree` component. This component is designed for efficiently constructing large Merkle sum trees by utilizing smaller-sized Merkle sum tree.
 
 ## Orchestrator
 
-The Orchestrator in the Summa Aggregation serves as the central management component, coordinating the data processing activities. It plays a pivotal role in coordinating the activities of `Executors` and `Workers`, improving of tasks in the generation of Merkle sum tree.
+The `Orchestrator` in the Summa Aggregation serves as the central management component, coordinating the data processing activities. It plays a pivotal role in coordinating the activities of `Executors` and `Workers`, improving of tasks in the generation of Merkle sum tree.
+The final result of the `Orchestrator` is the construction of the `AggregationMerkleSumTree`. This is achieved by aggregating the mini-trees constructed by the `Workers`.
+
 Key functions of the `Orchestrator` include:
 
 - **Dynamic Executor Spawning**: The Orchestrator dynamically spawns `Executors` in numbers set by the user. Each `Executor` is then connected to a dedicated `Worker` for efficient task execution.
@@ -15,9 +21,11 @@ Key functions of the `Orchestrator` include:
 
 - **Error Management and Pipeline Control**: The Orchestrator handles basic pipeline control and responds to errors by initiating the cancellation of all tasks.
 
+- **Build AggregationMerkleSumTree**: Its final result is that evaluate `AggregationMerkleSumTree` by aggregating the mini-trees generated by the `Workers`.
+
 ## Executor and Worker
 
-The `Executor` acts as a crucial intermediary between the `Orchestrator` and `Workers`, facilitating the data processing workflow. Spawned by the `Orchestrator`, each `Executor` operates in a one-to-one relationship with a `Worker`. The `Worker` is a server that can generate a merkle sum tree, internally called a `mini-tree`, by receiving entries data. The primary role of a `Worker` is to build these `mini-trees`, which are segments of the `AggregationMerkleSumTree`.
+The `Executor` acts as a crucial intermediary between the `Orchestrator` and `Workers`, facilitating the data processing workflow. Spawned by the `Orchestrator`, each `Executor` operates in a one-to-one relationship with a `Worker`. The `Worker` is a server that can generate a merkle sum tree, internally called a `mini-tree`, by receiving entries data. The primary role of a `Worker` is to build these `mini-tree`, which are segments of the `AggregationMerkleSumTree`.
 
 Key aspects of the `Executor's` role include:
 
@@ -25,9 +33,9 @@ Key aspects of the `Executor's` role include:
 
 - **Data Handling and Task Distribution**: A primary function of the `Executor` is to receive data entries, often parsed and prepared by the Orchestrator. Upon receiving these entries, the Executor is responsible for forwarding them to its connected `Worker`.
 
-- **Communication Bridge**: The `Executor` serves as a communication bridge within the data pipeline. It relays processed data, `mini-trees`, from Workers back to the Orchestrator.
+- **Communication Bridge**: The `Executor` serves as a communication bridge within the data pipeline. It relays processed data, `mini-tree`, from Workers back to the Orchestrator.
 
-The `Worker` called in here, mostly point out the container that runs `mini-tree-generator` server.
+The `Worker` called in here, mostly point out a container that runs `mini-tree-server`.
 
 ## ExecutorSpawner
 
@@ -65,9 +73,9 @@ You can initialize your Docker environment in Swarm mode, which is essential for
 
       This command will output information about the Swarm, including a join token.
 
-2. **Join Worker Nodes to the Swarm**:
+2. **Join Node to the Swarm**:
 
-      Use the join token provided by the main machine to add worker nodes to the swarm. On each worker node, run like:
+      Use the join token provided by the main machine to add nodes to the swarm. On each node, run like:
 
       ```bash
       Worker_1 $ docker swarm join --token <YOUR_JOIN_TOKEN> <MAIN_MACHINE_IP>:2377
@@ -75,7 +83,7 @@ You can initialize your Docker environment in Swarm mode, which is essential for
 
       Replace `<YOUR_JOIN_TOKEN>` with the actual token and `<MAIN_MACHINE_IP>` with the IP address of your main machine.
 
-3. **Verify Node Status**:
+3. **Check Node Status**:
   
       To confirm that the nodes are successfully joined to the swarm, check the node status on the main machine:
 
@@ -95,6 +103,8 @@ You can initialize your Docker environment in Swarm mode, which is essential for
       You are ready to spawn more workers!
 
 ### Spawning More Workers
+
+This section provides clear instructions on how to scale the `mini-tree-generator` service using Docker Compose and CLI commands.
 
 In Docker Swarm mode, containers are managed as services rather than by individual names. To spawn more workers, follow these steps:
 
@@ -128,8 +138,6 @@ In Docker Swarm mode, containers are managed as services rather than by individu
     This command shows the number of replicas running for each service in the swarm.
 
 Scaling allows you to adjust the number of service instances to meet your processing needs, enhancing the system's capability to handle increased loads or to improve redundancy and availability.
-
-This section provides clear instructions on how to scale the `mini-tree-generator` service using Docker Compose and CLI commands.
 
 ## Test
