--- src/orchestrator/orchestrator.rs
@@ -1,281 +1,0 @@
-use futures::future::join_all;
-use std::{cmp::min, error::Error};
-use tokio::sync::mpsc;
-use tokio_util::sync::CancellationToken;
-
-use super::entry_csv_parser::entry_parser;
-use crate::aggregation_merkle_sum_tree::AggregationMerkleSumTree;
-use crate::executor::ExecutorSpawner;
-
-pub struct Orchestrator<const N_ASSETS: usize, const N_BYTES: usize> {
-    executor_spawner: Box<dyn ExecutorSpawner>,
-    entry_csvs: Vec<String>,
-}
-
-impl<const N_ASSETS: usize, const N_BYTES: usize> Orchestrator<N_ASSETS, N_BYTES> {
-    fn new(executor_spawner: Box<dyn ExecutorSpawner>, entry_csvs: Vec<String>) -> Self {
-        Self {
-            executor_spawner,
-            entry_csvs,
-        }
-    }
-
-    // Calculate the range of tasks to be assigned to a executor.
-    //
-    // * `executor_index` - The index of the executor.
-    // * `total_executors` - The total number of executor.
-    //
-    // A tuple representing the start and end indices of the tasks assigned to the executor
-    fn calculate_task_range(
-        &self,
-        executor_index: usize,
-        total_executors: usize,
-    ) -> (usize, usize) {
-        let total_tasks = self.entry_csvs.len();
-        let base_tasks_per_executor = total_tasks / total_executors;
-        let extra_tasks = total_tasks % total_executors;
-
-        let start = executor_index * base_tasks_per_executor + min(executor_index, extra_tasks);
-        let end =
-            (executor_index + 1) * base_tasks_per_executor + min(executor_index + 1, extra_tasks);
-
-        (start, min(end, total_tasks))
-    }
-
-    /// Processes a list of CSV files concurrently using executors and aggregates the results.
-    ///
-    /// * `executor_count` - The number of executors to use.
-    ///
-    /// Data flow
-    ///
-    /// 1. Splits the list of CSV files into segments based on the number of available executors.
-    /// 2. A distribution thread loads each CSV file, parses it into `entries`, and sends these to `entries_tx`.
-    /// 3. Each executor receives `entries` from `entries_rx`, requests tasks to Worker, and sends results back through `tree_tx`.
-    /// 4. The processed data from all executors, collected from `tree_rx`, is aggregated into an `AggregationMerkleSumTree`.
-    /// 6. After processing, executors are terminated to release resources.
-    ///
-    async fn create_aggregation_mst(
-        self,
-        executor_count: usize,
-    ) -> Result<AggregationMerkleSumTree<N_ASSETS, N_BYTES>, Box<dyn Error>>
-    where
-        [usize; N_ASSETS + 1]: Sized,
-        [usize; 2 * (1 + N_ASSETS)]: Sized,
-    {
-        let entries_per_executor = self.entry_csvs.len() / executor_count;
-
-        let mut executors = Vec::new();
-        let mut result_collectors = Vec::new();
-
-        let channel_size = std::env::var("CHANNEL_SIZE")
-            .unwrap_or_default()
-            .parse::<usize>()
-            .unwrap_or(32);
-
-        let cancel_token = CancellationToken::new();
-        let actual_number_of_workers = min(executor_count, self.entry_csvs.len());
-        for i in 0..actual_number_of_workers {
-            // Declare channels for communication
-            //
-            // There are three channels are used inthis method.
-            //
-            // - A `entries_tx` receives parsed data from the entry parser to distribute tasks to executors.
-            // - A `tree_tx` channel is used by the executors to send the results of the tasks.
-            // - A `stop` channel is used to send a stop signal to all executors in case of an error,
-            //   allowing for a coordinated shutdown or cancellation of tasks.
-            //
-            let (entries_tx, mut entries_rx) = mpsc::channel(channel_size);
-            let (tree_tx, tree_rx) = mpsc::channel(channel_size);
-
-            let executor = self.executor_spawner.spawn_executor().await;
-            result_collectors.push((i, tree_rx));
-
-            let cloned_cancel_token = cancel_token.clone();
-            executors.push(tokio::spawn(async move {
-                        loop {
-                            tokio::select! {
-                                entries_data = entries_rx.recv() => {
-                                    // When the distribution thread is finished, the channel will be closed.
-                                    let entries = match entries_data {
-                                        Some(entries) => entries,
-                                        None => break,
-                                    };
-                                    let processed_task = match executor.generate_tree::<N_ASSETS, N_BYTES>(entries).await {
-                                        Ok(entries) => entries,
-                                        Err(e) => {
-                                            eprintln!("Executor_{:?}: error while processing entries {:?}", i, e);
-                                            cloned_cancel_token.cancel();
-                                            break;
-                                        }
-                                    };
-                                    if tree_tx.send(processed_task).await.is_err() {
-                                        eprintln!("Executor_{:?}: Error while sending tree result", i);
-                                        cloned_cancel_token.cancel();
-                                        break;
-                                    }
-                                },
-                                _ = cloned_cancel_token.cancelled() => {
-                                    eprintln!("Executor_{:?}: cancel signal received, terminating.", i);
-                                    break;
-                                },
-                            }
-                        }
-            }));
-
-            // 1. Distributing Tasks
-            //
-            // Spawn a distribution thread that distributes entries to executors
-            //
-            // - Loads CSV file from [csv_file_path].
-            // - Parses CSV file into 'entries'.
-            // - Sends 'entries' to executors via [entries_tx] channel.
-            //
-            let (start, end) = self.calculate_task_range(i, executor_count);
-            let entry_csvs_slice = self.entry_csvs[start..end].to_vec(); // Clone only the necessary slice
-
-            let cloned_cancel_token = cancel_token.clone();
-            tokio::spawn(async move {
-                for file_path in entry_csvs_slice.iter() {
-                    let entries = match entry_parser::<_, N_ASSETS, N_BYTES>(file_path) {
-                        Ok(entries) => entries,
-                        Err(e) => {
-                            eprintln!(
-                                "Executor_{:?}: Error while processing file {:?}: {:?}",
-                                i, file_path, e
-                            );
-                            cloned_cancel_token.cancel();
-                            break;
-                        }
-                    };
-
-                    tokio::select! {
-                        _ = cloned_cancel_token.cancelled() => {
-                            eprintln!("Executor_{:?}: cancel signal received, terminating distributor.", i);
-                            break;
-                        },
-                        send_entries = entries_tx.send(entries) => {
-                            if let Err(e) = send_entries {
-                                eprintln!("Executor_{:?}: Error while sending entries: {:?}", i, e);
-                                cloned_cancel_token.cancel();
-                                break;
-                            }
-                        }
-                    }
-                }
-                drop(entries_tx);
-            });
-        }
-
-        // 3. Collecting Results
-        //
-        // Collect `tree` results from executors
-        //
-        //  - Receives processed 'tree' from [tree_rx] channel.
-        //  - Collects all 'tree' results into 'all_tree_results'.
-        //  - Aggregates 'all_tree_results' into 'ordered_tree_results'.
-        //
-        let mut all_tree_responses = Vec::new();
-        for (index, mut tree_rx) in result_collectors {
-            let executor_results = tokio::spawn(async move {
-                let mut trees = Vec::new();
-                while let Some(result) = tree_rx.recv().await {
-                    trees.push(result);
-                }
-                (index, trees)
-            });
-            all_tree_responses.push(executor_results);
-        }
-
-        let all_tree_results = join_all(all_tree_responses).await;
-
-        // Aggregate results from all workers in order
-        let mut ordered_tree_results = vec![None; self.entry_csvs.len()];
-        for result in all_tree_results {
-            let (index, worker_results) = result.unwrap();
-            let start = index * entries_per_executor;
-            for (i, res) in worker_results.iter().enumerate() {
-                ordered_tree_results[start + i] = Some(res.clone());
-            }
-        }
-
-        // Terminate executors
-        self.executor_spawner.terminate_executors().await;
-
-        let all_merkle_sum_tree = ordered_tree_results.into_iter().flatten().collect();
-
-        AggregationMerkleSumTree::new(all_merkle_sum_tree)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::Orchestrator;
-    use crate::executor::{LocalSpawner, MockSpawner};
-    use summa_backend::merkle_sum_tree::Tree;
-
-    #[tokio::test]
-    async fn test_with_containers() {
-        let spawner = LocalSpawner::new(
-            "summa-aggregation".to_string(),
-            "orchestrator_test".to_string(),
-        );
-
-        let orchestrator = Orchestrator::<2, 14>::new(
-            Box::new(spawner),
-            vec![
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-            ],
-        );
-        let aggregation_merkle_sum_tree = orchestrator.create_aggregation_mst(2).await.unwrap();
-        assert_eq!(16, aggregation_merkle_sum_tree.mini_tree(0).entries.len());
-        assert_eq!(16, aggregation_merkle_sum_tree.mini_tree(1).entries.len());
-    }
-
-    #[tokio::test]
-    async fn test_single_mock_worker() {
-        let spawner = MockSpawner::new(None);
-
-        let orchestrator = Orchestrator::<2, 14>::new(
-            Box::new(spawner),
-            vec![
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-            ],
-        );
-        let aggregation_merkle_sum_tree = orchestrator.create_aggregation_mst(1).await.unwrap();
-        assert_eq!(16, aggregation_merkle_sum_tree.mini_tree(0).entries.len());
-        assert_eq!(16, aggregation_merkle_sum_tree.mini_tree(1).entries.len());
-    }
-
-    #[tokio::test]
-    async fn test_none_exist_csv() {
-        let spawner = MockSpawner::new(None);
-
-        let orchestrator = Orchestrator::<2, 14>::new(
-            Box::new(spawner),
-            vec![
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-                "./src/orchestrator/csv/no_exist.csv".to_string(),
-            ],
-        );
-        let one_mini_tree_result = orchestrator.create_aggregation_mst(2).await.unwrap();
-        assert_eq!(&0, one_mini_tree_result.depth());
-    }
-
-    #[tokio::test]
-    async fn test_none_exist_worker() {
-        let non_exist_worker_url = vec!["127.0.0.1:7878".to_string()];
-        let spawner = MockSpawner::new(Some(non_exist_worker_url));
-
-        let orchestrator = Orchestrator::<2, 14>::new(
-            Box::new(spawner),
-            vec![
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-                "./src/orchestrator/csv/entry_16.csv".to_string(),
-            ],
-        );
-        let empty_mini_tree_error = orchestrator.create_aggregation_mst(2).await.unwrap_err();
-        assert_eq!("Empty mini tree inputs", empty_mini_tree_error.to_string());
-    }
-}
